requirements.txt
langchain-core>=0.3.0
langchain-google-genai>=0.1.0
langchain-community>=0.0.30
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
